Packages/apps/Bluetooth core-stack fixes
    Merge bt.lnx.3.0 into bt.lnx 3.2
    GAP: Check object value when unRegister receiver

    Bluetooth: Fix bt profiles cannot be connected automatically
	           after un-bonding and bonding again

    BT: Remove profile from List in case of intent delay

    BT : Added a AdapterState check before sending start/stop
	     request to BT profiles

    GAP: Don't send discovery cancel when discovery not active

    Bluetooth: Fixed Static Analysis Issues
    - Made fixes for P1 KW issues

    Bluetooth: Added safe check before removing entry from ArrayList

    GAP: Avoid multiple locks inside getDevice method

    Enable the native wake lock in BT stack

    BT: Set foreground flag to start service intent

    BT: Sync objects lifecycle with life cycle of Service

    BT: Use local call to get BT State from BT Service

    BT: Dynamically load JNI library to avoid link error

    Bluetooth: packages/apps fixes porting on Android O

    Bluetooth: Handling the unhanded case to avoid failure

    Bluetooth: Load jni library before calling classinitnative

    Bluetooth: Replying more on AdapterState than Intent

    Opp: Using the eventHandler to start listener

    Bluetooth: Register new callback for HCI events callback

    Bluetooth: Do not send Pairing cancel intent during ACL disconnection

    Stop profile services when Bluetooth enable times out

    Broadcast UUIDS on BT state transition to ON

    GATT: Fix FD leak caused by GATT profile

    Clear disable flag during BT turn ON

    BT: Killing Bluetooth process in case of disable timeout

    Bluetooth: Gatt: Load jni library before calling classinitnative
    Fix to avoid the case when interface is already enabled.
    Bluetooth :Fix to avoid the access of JNI gloabal reference object
    Bluetooth: GAP: Update bondstatemachine to handle pairing scenarios
    Bluetooth: Null check added before accessing the Adapter service
    Bluetooth: Access to gloabal reference after it is deleted
    Bluetooth: Hedge against the Null global reference access
    Bluetooth: Stop the BT Adapter Service before stopping profiles
    Bluetooth: Prevent HF Exception in case of invalid telephoney context
    Bluetooth: Add NULL validation for Callback Objects
    Bluetooth: Acquire AdapterService handle in start of profile services
    Bluetooth: Fix for Intent to reach the receiver timebound
    Bluetooth: GAP: Acquire a wakelock to bring the display up.

    Bluetooth: Proper initialization of callback object for VS interface

    Bluetooth: Update the markings

    Bluetooth: Prevent unwanted discovery finished intent

    Used adapterService variable instead of mAdapterService

    BT App: Fix to resolve Static Code Analysis issues

    Bluetooth: Enable debug logs to debug app level issues

    Bluetooth: GAP: Added syncronization lock in Getdevice method

    Bluetooth: Introduced new HAL directory for VS interface

    Bluetooth: Including vendor HAL directory path for wearables

    GAP: Check prev discovery state during discovery operations

    GAP: Allow BT name upto 248 chars from setName

    BT: Make BondStateMachine while onCreate call of AdapterService

    BT: Calling disableProfileServices in case of BREDR_CLEANUP_TIMEOUT

    BT: Changed the handling of Process-Messages in timeout cases

    CRs-fixed: 2058416

+++ src/com/android/bluetooth/btservice/BondStateMachine.java
@@ -1,4 +1,6 @@
 /*
+ * Copyright (C) 2018 The Linux Foundation. All rights reserved.
+ * Not a Contribution
  * Copyright (C) 2012 The Android Open Source Project
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -25,6 +27,7 @@ import android.content.Intent;
 import android.os.Message;
 import android.os.UserHandle;
 import android.util.Log;
+import android.os.PowerManager;
 
 import com.android.bluetooth.Utils;
 import com.android.bluetooth.a2dp.A2dpService;
@@ -66,13 +69,20 @@ final class BondStateMachine extends StateMachine {
     private RemoteDevices mRemoteDevices;
     private BluetoothAdapter mAdapter;
 
+    /* The WakeLock is used for bringing up the LCD during a pairing request
+     * from remote device when Android is in Suspend state.*/
+    private PowerManager.WakeLock mWakeLock;
+
     private PendingCommandState mPendingCommandState = new PendingCommandState();
     private StableState mStableState = new StableState();
 
     public static final String OOBDATA = "oobdata";
 
-    private BondStateMachine(AdapterService service, AdapterProperties prop,
-            RemoteDevices remoteDevices) {
+    private final ArrayList<BluetoothDevice> mDevices =
+        new ArrayList<BluetoothDevice>();
+
+    private BondStateMachine(PowerManager pm, AdapterService service,
+            AdapterProperties prop, RemoteDevices remoteDevices) {
         super("BondStateMachine:");
         addState(mStableState);
         addState(mPendingCommandState);
@@ -81,12 +91,17 @@ final class BondStateMachine extends StateMachine {
         mAdapterProperties = prop;
         mAdapter = BluetoothAdapter.getDefaultAdapter();
         setInitialState(mStableState);
+
+        //WakeLock instantiation in RemoteDevices class
+        mWakeLock = pm.newWakeLock(PowerManager.FULL_WAKE_LOCK | PowerManager.ACQUIRE_CAUSES_WAKEUP
+                       | PowerManager.ON_AFTER_RELEASE, TAG);
+        mWakeLock.setReferenceCounted(false);
     }
 
-    public static BondStateMachine make(AdapterService service, AdapterProperties prop,
-            RemoteDevices remoteDevices) {
+    public static BondStateMachine make(PowerManager pm, AdapterService service,
+            AdapterProperties prop, RemoteDevices remoteDevices) {
         Log.d(TAG, "make");
-        BondStateMachine bsm = new BondStateMachine(service, prop, remoteDevices);
+        BondStateMachine bsm = new BondStateMachine(pm, service, prop, remoteDevices);
         bsm.start();
         return bsm;
     }
@@ -132,12 +147,15 @@ final class BondStateMachine extends StateMachine {
                     break;
                 case BONDING_STATE_CHANGE:
                     int newState = msg.arg1;
-                /* if incoming pairing, transition to pending state */
+                    /* if incoming pairing, transition to pending state */
                     if (newState == BluetoothDevice.BOND_BONDING) {
+                        if (!mDevices.contains(dev)) {
+                            mDevices.add(dev);
+                        }
                         sendIntent(dev, newState, 0);
                         transitionTo(mPendingCommandState);
                     } else if (newState == BluetoothDevice.BOND_NONE) {
-                    /* if the link key was deleted by the stack */
+                        /* if the link key was deleted by the stack */
                         sendIntent(dev, newState, 0);
                     } else {
                         Log.e(TAG, "In stable state, received invalid newState: " + newState);
@@ -155,7 +173,6 @@ final class BondStateMachine extends StateMachine {
 
 
     private class PendingCommandState extends State {
-        private final ArrayList<BluetoothDevice> mDevices = new ArrayList<BluetoothDevice>();
 
         @Override
         public void enter() {
@@ -195,6 +212,14 @@ final class BondStateMachine extends StateMachine {
                     int reason = getUnbondReasonFromHALCode(msg.arg2);
                     sendIntent(dev, newState, reason);
                     if (newState != BluetoothDevice.BOND_BONDING) {
+                        // check if bond none is received from device which
+                        // was in pairing state otherwise don't transition to
+                        // stable state.
+                        if (newState == BluetoothDevice.BOND_NONE &&
+                            !mDevices.contains(dev) && mDevices.size() != 0) {
+                            infoLog("not transitioning to stable state");
+                            break;
+                        }
                         /* this is either none/bonded, remove and transition */
                         result = !mDevices.remove(dev);
                         if (mDevices.isEmpty()) {
@@ -222,11 +247,21 @@ final class BondStateMachine extends StateMachine {
                 case SSP_REQUEST:
                     int passkey = msg.arg1;
                     int variant = msg.arg2;
+                    if (devProp == null)
+                    {
+                        Log.e(TAG,"Received msg from an unknown device");
+                        return false;
+                    }
                     sendDisplayPinIntent(devProp.getAddress(), passkey, variant);
                     break;
                 case PIN_REQUEST:
                     BluetoothClass btClass = dev.getBluetoothClass();
                     int btDeviceClass = btClass.getDeviceClass();
+                    if (devProp == null)
+                    {
+                        Log.e(TAG,"Received msg from an unknown device");
+                        return false;
+                    }
                     if (btDeviceClass == BluetoothClass.Device.PERIPHERAL_KEYBOARD || btDeviceClass
                             == BluetoothClass.Device.PERIPHERAL_KEYBOARD_POINTING) {
                         // Its a keyboard. Follow the HID spec recommendation of creating the
@@ -266,6 +301,7 @@ final class BondStateMachine extends StateMachine {
     }
 
     private boolean cancelBond(BluetoothDevice dev) {
+        if (mAdapterService == null) return false;
         if (dev.getBondState() == BluetoothDevice.BOND_BONDING) {
             byte[] addr = Utils.getBytesFromAddress(dev.getAddress());
             if (!mAdapterService.cancelBondNative(addr)) {
@@ -278,6 +314,7 @@ final class BondStateMachine extends StateMachine {
     }
 
     private boolean removeBond(BluetoothDevice dev, boolean transition) {
+        if (mAdapterService == null) return false;
         if (dev.getBondState() == BluetoothDevice.BOND_BONDED) {
             byte[] addr = Utils.getBytesFromAddress(dev.getAddress());
             if (!mAdapterService.removeBondNative(addr)) {
@@ -295,6 +332,7 @@ final class BondStateMachine extends StateMachine {
 
     private boolean createBond(BluetoothDevice dev, int transport, OobData oobData,
             boolean transition) {
+        if (mAdapterService == null) return false;
         if (dev.getBondState() == BluetoothDevice.BOND_NONE) {
             infoLog("Bond address is:" + dev);
             byte[] addr = Utils.getBytesFromAddress(dev.getAddress());
@@ -317,6 +355,9 @@ final class BondStateMachine extends StateMachine {
     }
 
     private void sendDisplayPinIntent(byte[] address, int pin, int variant) {
+
+        // Acquire wakelock during PIN code request to bring up LCD display
+        mWakeLock.acquire();
         Intent intent = new Intent(BluetoothDevice.ACTION_PAIRING_REQUEST);
         intent.putExtra(BluetoothDevice.EXTRA_DEVICE, mRemoteDevices.getDevice(address));
         if (pin != 0) {
@@ -327,6 +368,8 @@ final class BondStateMachine extends StateMachine {
         // Workaround for Android Auto until pre-accepting pairing requests is added.
         intent.addFlags(Intent.FLAG_RECEIVER_INCLUDE_BACKGROUND);
         mAdapterService.sendOrderedBroadcast(intent, mAdapterService.BLUETOOTH_ADMIN_PERM);
+        // Release wakelock to allow the LCD to go off after the PIN popup notification.
+        mWakeLock.release();
     }
 
     private void sendIntent(BluetoothDevice device, int newState, int reason) {
